django/requirements/base.txt
Django>=1.8.4
django-environ>=0.3.0
django-braces>=1.8.0
django-crispy-forms>=1.4.0
django-admin-bootstrapped>=2.4.0
django-authtools>=1.2.0
djangorestframework>=3.2.3
wheel==0.24.0
drf-nested-routers>=0.10.0
